Create astropy.stats.lombscargle.statistics; move pdf and cdf functions there
