+++ astropy/stats/lombscargle/core.py
@@ -356,105 +356,3 @@ class LombScargle(object):
                              fit_mean=self.fit_mean,
                              nterms=self.nterms)
         return y_fit * get_unit(self.y)
-
-    def cumulative_distribution(self, z, normalization='standard'):
-        """
-        Compute the theoretical cumulative distribution of the periodogram
-        for the null hypothesis - i.e. data consisting of Gaussian noise.
-
-        Parameters
-        ----------
-        z : array-like
-            the periodogram value
-        normalziation : string
-            The periodogram normalization. Must be one of
-            ['standard', 'model', 'log', 'psd']
-
-        Returns
-        -------
-        cdf : np.ndarray
-            The theoretical cumulative distribution function
-
-        Notes
-        -----
-        Expressions used here are adapted from Baluev 2008, Table 1.
-        """
-        if isinstance(z, units.Quantity):
-            if z.unit == units.dimensionless_unscaled:
-                z = z.value
-            else:
-                raise ValueError('The distribution can be computed only for '
-                                 'normalized (unitless) periodograms.')
-
-        N = len(self.t)
-        if self.nterms != 1:
-            raise NotImplementedError("distribution is not implemented for"
-                                      "the multi-term periodogram")
-
-        if self.fit_mean or self.center_data:
-            Nk = N - 3
-        else:
-            Nk = N - 2
-
-        if normalization == 'psd':
-            return 1 - np.exp(-z)
-        elif normalization == 'standard':
-            return 1 - (1 + z) ** (-0.5 * Nk)
-        elif normalization == 'model':
-            return 1 - (1 - z) ** (0.5 * Nk)
-        elif normalization == 'log':
-            return 1 - np.exp(-0.5 * Nk * z)
-        else:
-            raise ValueError("normalization='{0}' is not recognized"
-                             "".format(normalization))
-
-    def distribution(self, z, normalization='standard'):
-        """
-        Compute the theoretical probability density function of the periodogram
-        for the null hypothesis - i.e. data consisting of Gaussian noise.
-
-        Parameters
-        ----------
-        z : array-like
-            the periodogram value
-        normalziation : string
-            The periodogram normalization. Must be one of
-            ['standard', 'model', 'log', 'psd']
-
-        Returns
-        -------
-        pdf : np.ndarray
-            The theoretical probability density function
-
-        Notes
-        -----
-        Expressions used here are adapted from Baluev 2008, Table 1.
-        """
-        if isinstance(z, units.Quantity):
-            if z.unit == units.dimensionless_unscaled:
-                z = z.value
-            else:
-                raise ValueError('The distribution can be computed only for '
-                                 'normalized (unitless) periodograms.')
-
-        N = len(self.t)
-        if self.nterms != 1:
-            raise NotImplementedError("distribution is not implemented for"
-                                      "the multi-term periodogram")
-
-        if self.fit_mean or self.center_data:
-            Nk = N - 3
-        else:
-            Nk = N - 2
-
-        if normalization == 'psd':
-            return np.exp(-z)
-        elif normalization == 'standard':
-            return 0.5 * Nk * (1 + z) ** (-0.5 * Nk - 1)
-        elif normalization == 'model':
-            return 0.5 * Nk * (1 - z) ** (0.5 * Nk - 1)
-        elif normalization == 'log':
-            return 0.5 * Nk * np.exp(-0.5 * Nk * z)
-        else:
-            raise ValueError("normalization='{0}' is not recognized"
-                             "".format(normalization))
